crypto_ai_bot/requirements.txt
pandas
pandas_ta
python-binance
pyyaml